+++ the-graph/the-graph-graph.js
@@ -94,7 +94,9 @@
         selectedNodes: [],
         errorNodes: [],
         selectedEdges: [],
-        animatedEdges: []
+        animatedEdges: [],
+        offsetX: this.props.offsetX,
+        offsetY: this.props.offsetY
       };
     },
     componentDidMount: function () {
@@ -110,7 +112,7 @@
       this.props.graph.on("changeOutport", this.markDirty);
       this.props.graph.on("endTransaction", this.markDirty);
 
-      this.getDOMNode().addEventListener("the-graph-node-remove", this.removeNode);
+      ReactDOM.findDOMNode(this).addEventListener("the-graph-node-remove", this.removeNode);
     },
     edgePreview: null,
     edgeStart: function (event) {
@@ -143,7 +145,7 @@
       edge.metadata = { route: event.detail.route };
       edge.type = event.detail.port.type;
 
-      var appDomNode = this.props.app.getDOMNode();
+      var appDomNode = ReactDOM.findDOMNode(this.props.app);
       appDomNode.addEventListener("mousemove", this.renderPreviewEdge);
       appDomNode.addEventListener("track", this.renderPreviewEdge);
       // TODO tap to add new node here
@@ -152,7 +154,7 @@
       this.setState({edgePreview: edge});
     },
     cancelPreviewEdge: function (event) {
-      var appDomNode = this.props.app.getDOMNode();
+      var appDomNode = ReactDOM.findDOMNode(this.props.app);
       appDomNode.removeEventListener("mousemove", this.renderPreviewEdge);
       appDomNode.removeEventListener("track", this.renderPreviewEdge);
       appDomNode.removeEventListener("tap", this.cancelPreviewEdge);
@@ -164,6 +166,8 @@
     renderPreviewEdge: function (event) {
       var x = event.x || event.clientX || 0;
       var y = event.y || event.clientY || 0;
+      x -= this.props.app.state.offsetX || 0;
+      y -= this.props.app.state.offsetY || 0;
       var scale = this.props.app.state.scale;
       this.setState({
         edgePreviewX: (x - this.props.app.state.x) / scale,