+++ the-graph/the-graph-menu.js
@@ -156,20 +156,20 @@
     },
     componentDidMount: function () {
       if (this.state.n4tappable) {
-        this.refs.n4.getDOMNode().addEventListener("up", this.onTapN4);
+        this.refs.n4.addEventListener("up", this.onTapN4);
       }
       if (this.state.s4tappable) {
-        this.refs.s4.getDOMNode().addEventListener("up", this.onTapS4);
+        this.refs.s4.addEventListener("up", this.onTapS4);
       }
       if (this.state.e4tappable) {
-        this.refs.e4.getDOMNode().addEventListener("up", this.onTapE4);
+        this.refs.e4.addEventListener("up", this.onTapE4);
       }
       if (this.state.w4tappable) {
-        this.refs.w4.getDOMNode().addEventListener("up", this.onTapW4);
+        this.refs.w4.addEventListener("up", this.onTapW4);
       }
 
       // Prevent context menu
-      this.getDOMNode().addEventListener("contextmenu", function (event) {
+      ReactDOM.findDOMNode(this).addEventListener("contextmenu", function (event) {
         if (event) {
           event.stopPropagation();
           event.preventDefault();
@@ -261,8 +261,8 @@
   TheGraph.ModalBG = React.createFactory( React.createClass({
     displayName: "TheGraphModalBG",
     componentDidMount: function () {
-      var domNode = this.getDOMNode();
-      var rectNode = this.refs.rect.getDOMNode();
+      var domNode = ReactDOM.findDOMNode(this);
+      var rectNode = this.refs.rect; 
 
       // Right-click on another item will show its menu
       domNode.addEventListener("down", function (event) {